refactor(example): modularize project architecture and enhance NestLens integration
- Replaced single-file 'main.ts' with a modular NestJS structure.
- Added Swagger API documentation for better observability.
