--- example/src/CLAUDE.md
@@ -1,11 +1,0 @@
-<claude-mem-context>
-# Recent Activity
-
-<!-- This section is auto-generated by claude-mem. Edit content outside the tags. -->
-
-### Jan 8, 2026
-
-| ID | Time | T | Title | Read |
-|----|------|---|-------|------|
-| #8 | 12:48 AM | 🔴 | Fixed GraphQL watcher dependency injection issue in production app | ~508 |
-</claude-mem-context>
