+++ example/src/posts/entities/comment.entity.ts
@@ -1,0 +1,10 @@
+import { ObjectType, Field, ID } from '@nestjs/graphql';
+
+@ObjectType()
+export class Comment {
+    @Field(() => ID)
+    id!: string;
+
+    @Field()
+    text!: string;
+}
