+++ example/src/products/dto/update-product.input.ts
@@ -1,0 +1,10 @@
+import { InputType, Field, Float } from '@nestjs/graphql';
+
+@InputType()
+export class UpdateProductInput {
+    @Field({ nullable: true })
+    name?: string;
+
+    @Field(() => Float, { nullable: true })
+    price?: number;
+}
