+++ example/src/products/products.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ProductsResolver } from './products.resolver';
+import { DatabaseModule } from '../database/database.module';
+
+@Module({
+  imports: [DatabaseModule],
+  providers: [ProductsResolver],
+})
+export class ProductsModule { }
