+++ example/src/status/status.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { StatusController, ErrorController } from './status.controller';
+
+@Module({
+    controllers: [StatusController, ErrorController],
+})
+export class StatusModule { }
