+++ example/src/users/dto/create-user.input.ts
@@ -1,0 +1,10 @@
+import { InputType, Field } from '@nestjs/graphql';
+
+@InputType()
+export class CreateUserInput {
+    @Field()
+    name!: string;
+
+    @Field()
+    email!: string;
+}
